Moved helper methods out of config file
