+++ data-prepper-plugins/translate-processor/src/main/java/org/opensearch/dataprepper/plugins/processor/translate/TargetsParameterConfig.java
@@ -5,9 +5,6 @@ import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import org.apache.commons.lang3.Range;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.math.NumberUtils;
-import org.opensearch.dataprepper.model.plugin.InvalidPluginConfigurationException;
 import org.opensearch.dataprepper.plugins.processor.mutateevent.TargetType;
 import org.opensearch.dataprepper.typeconverter.TypeConverter;
 
@@ -40,7 +37,9 @@ public class TargetsParameterConfig {
     private TargetType targetType = TargetType.STRING;
 
     public TargetsParameterConfig(Map<String, Object> map, String target, RegexParameterConfiguration regexParameterConfig, String translateWhen, String defaultValue, TargetType targetType) {
-        this.targetType = Optional.ofNullable(targetType).orElse(TargetType.STRING);
+        this.targetType = Optional
+                .ofNullable(targetType)
+                .orElse(TargetType.STRING);
         this.target = target;
         this.map = map;
         this.defaultValue = defaultValue;
@@ -50,16 +49,12 @@ public class TargetsParameterConfig {
         parseMappings();
     }
 
-    public void parseMappings() {
-        if (Objects.nonNull(getRegexParameterConfiguration())) {
-            compilePatterns(getRegexParameterConfiguration().getPatterns());
-        }
-        processMapField(map);
-        checkOverlappingKeys();
-    }
-
     public String getTarget() {
         return target;
+    }
+
+    public Map<String, Object> getMap() {
+        return map;
     }
 
     public String getDefaultValue() {
@@ -132,73 +127,11 @@ public class TargetsParameterConfig {
         return true;
     }
 
-    private void compilePatterns(Map<String, Object> mappings) {
-        for (String pattern : mappings.keySet()) {
-            Pattern compiledPattern = Pattern.compile(pattern);
-            compiledPatterns.put(compiledPattern, mappings.get(pattern));
-        }
-    }
-
-    private void processMapField(Map<String, Object> map) {
-        if (Objects.nonNull(map)) {
-            for (Map.Entry<String, Object> mapEntry : map.entrySet()) {
-                parseIndividualKeys(mapEntry);
-            }
-        }
-    }
-
-    private void parseIndividualKeys(Map.Entry<String, Object> mapEntry) {
-        String[] commaSeparatedKeys = mapEntry.getKey().split(",");
-        for (String individualKey : commaSeparatedKeys) {
-            if (individualKey.contains("-")) {
-                addRangeMapping(Map.entry(individualKey, mapEntry.getValue()));
-            } else {
-                addIndividualMapping(individualKey, mapEntry.getValue());
-            }
-        }
-    }
-
-    private void addRangeMapping(Map.Entry<String, Object> mapEntry) {
-        String[] rangeKeys = mapEntry.getKey().split("-");
-        if (rangeKeys.length != 2 || !StringUtils.isNumericSpace(rangeKeys[0]) || !StringUtils.isNumericSpace(rangeKeys[1])) {
-            addIndividualMapping(mapEntry.getKey(), mapEntry.getValue());
-        } else {
-            Float lowKey = Float.parseFloat(rangeKeys[0]);
-            Float highKey = Float.parseFloat(rangeKeys[1]);
-            Range<Float> rangeEntry = Range.between(lowKey, highKey);
-            if (isRangeOverlapping(rangeEntry)) {
-                String exceptionMsg = "map option contains key " + mapEntry.getKey() + " that overlaps with other range entries";
-                throw new InvalidPluginConfigurationException(exceptionMsg);
-            } else {
-                rangeMappings.put(Range.between(lowKey, highKey), mapEntry.getValue());
-            }
-        }
-    }
-
-    private void addIndividualMapping(final String key, final Object value) {
-        if (individualMappings.containsKey(key)) {
-            String exceptionMsg = "map option contains duplicate entries of " + key;
-            throw new InvalidPluginConfigurationException(exceptionMsg);
-        } else {
-            individualMappings.put(key.strip(), value);
-        }
-    }
-
-    private boolean isRangeOverlapping(Range<Float> rangeEntry) {
-        return rangeMappings.keySet().stream().anyMatch(range -> range.isOverlappedBy(rangeEntry));
-    }
-
-    private void checkOverlappingKeys() {
-        for (String individualKey : individualMappings.keySet()) {
-            if (NumberUtils.isParsable(individualKey)) {
-                Float floatKey = Float.parseFloat(individualKey);
-                Range<Float> range = Range.between(floatKey, floatKey);
-                if (isRangeOverlapping(range)) {
-                    String exceptionMsg = "map option contains key " + individualKey + " that overlaps with other range entries";
-                    throw new InvalidPluginConfigurationException(exceptionMsg);
-                }
-            }
-        }
+    public void parseMappings() {
+        MappingsParser parser = new MappingsParser(this);
+        individualMappings.putAll(parser.fetchIndividualMappings());
+        rangeMappings.putAll(parser.fetchRangeMappings());
+        compiledPatterns.putAll(parser.fetchCompiledPatterns());
     }
 
 }
